consolidate windows_service script
