+++ .gitattributes
@@ -1,3 +1,3 @@
 *.bat text eol=crlf
 *.ps1 text eol=crlf
-windows_service.py text eol=crlf
+tools/windows_service.py text eol=crlf
